PM-55 - Move PreloginRequest model into auth requests model folder as it should be + update refs.
