+++ libs/common/src/auth/abstractions/identity-api.service.ts
@@ -1,7 +1,7 @@
-import { PreloginRequest } from "../../models/request/prelogin.request";
 import { PasswordTokenRequest } from "../models/request/identity-token/password-token.request";
 import { SsoTokenRequest } from "../models/request/identity-token/sso-token.request";
 import { UserApiTokenRequest } from "../models/request/identity-token/user-api-token.request";
+import { PreloginRequest } from "../models/request/prelogin.request";
 import { IdentityCaptchaResponse } from "../models/response/identity-captcha.response";
 import { IdentityTokenResponse } from "../models/response/identity-token.response";
 import { IdentityTwoFactorResponse } from "../models/response/identity-two-factor.response";
