+++ libs/common/src/auth/models/request/prelogin.request.ts
@@ -1,4 +1,3 @@
-// TODO: Should this be moved into auth request folder?
 export class PreloginRequest {
   email: string;
