+++ libs/common/src/auth/services/auth.service.ts
@@ -14,7 +14,6 @@ import { KdfType } from "../../enums/kdfType";
 import { KeySuffixOptions } from "../../enums/keySuffixOptions";
 import { Utils } from "../../misc/utils";
 import { SymmetricCryptoKey } from "../../models/domain/symmetric-crypto-key";
-import { PreloginRequest } from "../../models/request/prelogin.request";
 import { ErrorResponse } from "../../models/response/error.response";
 import { AuthRequestPushNotification } from "../../models/response/notification.response";
 import { AuthService as AuthServiceAbstraction } from "../abstractions/auth.service";
@@ -38,6 +37,7 @@ import {
 } from "../models/domain/log-in-credentials";
 import { TokenTwoFactorRequest } from "../models/request/identity-token/token-two-factor.request";
 import { PasswordlessAuthRequest } from "../models/request/passwordless-auth.request";
+import { PreloginRequest } from "../models/request/prelogin.request";
 import { AuthRequestResponse } from "../models/response/auth-request.response";
 
 const sessionTimeoutLength = 2 * 60 * 1000; // 2 minutes
